messages: separate "help" links from normal links
Some links in the `MessagesView` are part of Element X itself,
such as the help pages in `LearnMoreConfig` [1]. These links
should open in an "in-app" Chrome Custom Tab, because they are
basically part of the app.

Web links from chat messages, on the other hand, should open in
the user's regular preferred web browser as regular tabs.

Separate "regular" links from "help" links with a different
callback for each:

* `onLinkClick` is modified to use regular browser tabs

* `onHelpLinkClick` prefers a Chrome Custom Tab where possible

[1]: appconfig/src/main/kotlin/io/element/android/appconfig/LearnMoreConfig.kt